+++ azure-monitor-opentelemetry-distro/azure/monitor/opentelemetry/distro/__init__.py
@@ -17,21 +17,13 @@ from azure.monitor.opentelemetry.exporter import (
 )
 
 
-_logger = logging.getLogger(__name__)
-
-
 def configure_opentelemetry(**kwargs):
-    # pylint: disable=too-many-locals
-    # pylint: disable=too-many-statements
     """
-    Configures OpenTelemetry with Lightstep environment variables
-    This function works as a configuration layer that allows the Lightstep end
-    user to use current environment variables seamlessly with OpenTelemetry. In
-    this way, it is not needed to make any configuration changes to the
-    environment before using OpenTelemetry. The configuration can be done via
-    environment variables (prefixed with `LS`) or via arguments passed to this
-    function. Each argument has a 1:1 correspondence with an environment
-    variable, their description follows:
+    This function works as a configuration layer that allows the
+    end user to configure OpenTelemetry and Azure monitor components. The
+    configuration can be done via environment variables or
+    via arguments passed to this function. Each argument has a 1:1
+    correspondence with an environment variable.
     """
 
     configurations = get_configurations(**kwargs)
@@ -39,7 +31,6 @@ def configure_opentelemetry(**kwargs):
     service_name = configurations["service_name"]
     service_namespace = configurations["service_namespace"]
     service_instance_id = configurations["service_instance_id"]
-    service_version = configurations["service_version"]
     disable_tracing = configurations["disable_tracing"]
 
     if not disable_tracing:
@@ -47,7 +38,6 @@ def configure_opentelemetry(**kwargs):
                 ResourceAttributes.SERVICE_NAME: service_name,
                 ResourceAttributes.SERVICE_NAMESPACE: service_namespace,
                 ResourceAttributes.SERVICE_INSTANCE_ID: service_instance_id,
-                ResourceAttributes.SERVICE_VERSION: service_version,
             }
         )
         trace.set_tracer_provider(TracerProvider(resource=resource))